chore(lib): move common traits to types.rs
