+++ lspresso-shot/init_dot_lua.rs
@@ -115,7 +115,7 @@ fn get_attach_action(test_type: TestType) -> String {
         TestType::Declaration => include_str!("lua_templates/declaration_action.lua"),
         TestType::Definition => include_str!("lua_templates/definition_action.lua"),
         TestType::Diagnostic => "\n-- NOTE: No `check_progress_result` function for diagnostics, instead handled by `DiagnosticChanged` autocmd\n",
-        TestType::DocumentSymbol => include_str!("lua_templates/document_symbol.lua"),
+        TestType::DocumentHighlight => include_str!("lua_templates/document_highlight_action.lua"),
         TestType::DocumentSymbol => include_str!("lua_templates/document_symbol_action.lua"),
         TestType::Formatting => include_str!("lua_templates/formatting_action.lua"),
         TestType::Hover => include_str!("lua_templates/hover_action.lua"),
